+++ include/CExprIValue.h
@@ -11,9 +11,9 @@ class CExprIntegerValue : public CExprValueBase {
     return new CExprIntegerValue(integer_);
   }
 
-  bool getBooleanValue(bool        &b) const { b = (integer_ != 0); return true; }
-  bool getIntegerValue(long        &l) const { l = integer_       ; return true; }
-  bool getRealValue   (double      &r) const { r = integer_       ; return true; }
+  bool getBooleanValue(bool        &b) const { b = (integer_ != 0) ; return true; }
+  bool getIntegerValue(long        &l) const { l = integer_        ; return true; }
+  bool getRealValue   (double      &r) const { r = double(integer_); return true; }
   bool getStringValue (std::string &s) const;
 
   void setIntegerValue(long l) { integer_ = l; }
