+++ include/CExprInterp.h
@@ -31,7 +31,7 @@ class CExprIToken {
   const std::string &getString    () const { return base()->getString    (); }
 
   uint getNumChildren() const {
-    return children_.size();
+    return uint(children_.size());
   }
 
   CExprITokenPtr getChild(uint i) {
@@ -86,7 +86,7 @@ class CExprITokenStack {
   }
 
   uint getNumTokens() const {
-    return stack_.size();
+    return uint(stack_.size());
   }
 
   CExprITokenPtr getToken(uint i) const {
